Updated Models and Session Id Support
1) Support SessionId login along with apikey/tokens.
2) Updated models with 6.8.1 version
3) Support for Oracle Restore

+++ src/Cohesity.Powershell.Models/DatastoreInfo.cs
@@ -1,6 +1,5 @@
 // Copyright 2019 Cohesity Inc.
 
-
 using System;
 using System.Linq;
 using System.IO;
@@ -12,7 +11,6 @@ using System.Collections.ObjectModel;
 using System.Runtime.Serialization;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
-
 
 namespace Cohesity.Model
 {
@@ -27,7 +25,7 @@ namespace Cohesity.Model
         /// </summary>
         /// <param name="capacity">Specifies the capacity of the datastore in bytes..</param>
         /// <param name="freeSpace">Specifies the available space on the datastore in bytes..</param>
-        public DatastoreInfo(ulong? capacity = default(ulong?), ulong? freeSpace = default(ulong?))
+        public DatastoreInfo(int? capacity = default(int?), int? freeSpace = default(int?))
         {
             this.Capacity = capacity;
             this.FreeSpace = freeSpace;
@@ -40,14 +38,14 @@ namespace Cohesity.Model
         /// </summary>
         /// <value>Specifies the capacity of the datastore in bytes.</value>
         [DataMember(Name="capacity", EmitDefaultValue=true)]
-        public ulong? Capacity { get; set; }
+        public int? Capacity { get; set; }
 
         /// <summary>
         /// Specifies the available space on the datastore in bytes.
         /// </summary>
         /// <value>Specifies the available space on the datastore in bytes.</value>
         [DataMember(Name="freeSpace", EmitDefaultValue=true)]
-        public ulong? FreeSpace { get; set; }
+        public int? FreeSpace { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
